# api/openapi/control-plane.yaml
openapi: 3.1.0
info:
  title: Otherix Control Plane API
  version: "0.1.0"
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    Otherix Control Plane REST API. Source of truth for server stubs and
    client SDK generation.

    Coverage so far:
      - Part 1 — auth, users, nodes, join-tokens.
      - Part 2 — firmwares, storage-pools, networks, vms, tasks.

    User-owned resources carry an `owner_id` (multi-tenancy is not
    modelled).
servers:
  - url: https://example.invalid/api
    description: Placeholder; deployment URL substituted by client at runtime.
...
paths:
...
  /v1/nodes/{name}/heartbeat:
    parameters:
      - name: name
        in: path
        required: true
        schema: { type: string, maxLength: 253 }
        description: |
          Cluster-unique node name. The agent identifies itself by the
          name encoded in its cert CN (`node-<name>`); UUIDs are never
          used on the heartbeat path. Agent identity is name-keyed.
    post:
      operationId: agents.heartbeat
      tags: [agents]
      summary: Receive a periodic state snapshot from a node agent.
      description: |
        Called by the node agent on a periodic cadence (default 30 s,
        operator-tunable). Carries a compact snapshot of node-level
        capabilities, current resource availability, and a per-VM
        runtime report.

        On receipt the Control Plane updates `nodes.last_heartbeat_at`,
        upserts capabilities into the matching `nodes` columns and the
        `nodes.capabilities` jsonb, refreshes the `node_firmwares`
        rows, and upserts each `vm_runtime` row from `vms[]`. A
        divergence between the reported `vms[]` and the CP's desired
        state schedules reconciliation work — that is internal CP
        logic, not part of the contract surface.

        Idempotency: this endpoint **does not** accept
        `Idempotency-Key`. Heartbeats are a continuous stream of
        snapshots; repeating the same body harms nothing, and the
        response carries no value to cache. Same carve-out logic as
        `auth.login` / `auth.refresh`.

        Future-compatibility: the response is intentionally minimal
        today. Later versions may include directives for the agent
        (e.g. requested re-scans, configuration nudges); clients must
        accept unknown response fields gracefully.
      security:
        - agentMTLS: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HeartbeatRequest"
      responses:
        "200":
          description: Heartbeat accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HeartbeatResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401":
          description: |
            Client certificate missing, untrusted, or its SAN does not
            match a known agent_cert. Distinct from the user-JWT
            `Unauthenticated` response — `details.reason` is one of
            `cert_untrusted` / `cert_san_unknown` / `cert_revoked`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Node name not found (or soft-deleted).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: |
            Node-level conflict. `details.reason` is one of:
              - `architecture_mismatch` — heartbeat reports an
                architecture that differs from the registered
                `nodes.architecture`. Architecture is immutable; the
                agent must re-register through the join-token flow if
                the underlying hardware changed.
              - `node_gone` — node is in `gone` status; further
                heartbeats are rejected until the operator
                un-deletes it (or it is re-registered).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500": { $ref: "#/components/responses/Internal" }
...
  /v1/storage-pools:
    get:
      operationId: storagePools.list
      tags: [storage-pools]
      summary: List storage pools.
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/NodeIdFilter"
        - name: type
          in: query
          required: false
          schema: { type: string, enum: [local_dir] }
      responses:
        "200":
          description: List of storage pools.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoragePoolList"
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "500": { $ref: "#/components/responses/Internal" }
    post:
      operationId: storagePools.create
      tags: [storage-pools]
      summary: Create a storage pool on a node (admin only).
      description: |
        **Required permission:** `storage_pool:manage` (admin).
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoragePoolCreate"
      responses:
        "201":
          description: Storage pool created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoragePool"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/PermissionDenied" }
        "404":
          description: Referenced node missing or invisible to caller.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/Internal" }
...
  /v1/vms:
    get:
      operationId: vms.list
      tags: [vms]
      summary: List virtual machines.
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/ArchitectureFilter"
        - $ref: "#/components/parameters/OwnerIdFilter"
        - $ref: "#/components/parameters/NodeIdFilter"
        - name: desired_phase
          in: query
          required: false
          schema:
            type: string
            enum: [running, stopped, deleted]
          description: Filter by desired phase (the user-set state).
        - name: name
          in: query
          required: false
          schema: { type: string }
          description: Filter by name substring (case-insensitive).
      responses:
        "200":
          description: |
            List of VMs. List items omit the embedded `disks` and `nics`
            collections — fetch a single VM with `?include=disks,nics` to
            retrieve them.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VMList"
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "500": { $ref: "#/components/responses/Internal" }
    post:
      operationId: vms.create
      tags: [vms]
      summary: Create a virtual machine.
      description: |
        VM creation is asynchronous: the request returns 202 with a task
        id; the VM row exists immediately in `desired_phase = stopped`,
        but agent-side allocation (image cache, disk preallocation, QEMU
        process if `desired_phase = running`) happens in the background.
        Poll the task to observe completion.

        The image source is supplied directly on the body (`image_url`
        + optional `image_sha256` content pin + `format`); the VM row is
        self-describing and there is no template entity. A single root
        disk is provisioned from the image, sized to `disk_gib` (or the
        image's virtual size when `disk_gib` is 0/omitted).
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VMCreate"
      responses:
        "202":
          description: VM creation accepted; poll the returned task.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AsyncTaskAccepted"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/PermissionDenied" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/Internal" }
...
  /v1/vms/{id}/console:
    parameters:
      - $ref: "#/components/parameters/VMName"
    post:
      operationId: vms.console
      tags: [vms]
      summary: Issue a short-lived console-session token (synchronous).
      description: |
        Returns a single-use token, a websocket URL, and an expiry
        for opening a console session against the VM's running QEMU
        process. The websocket protocol itself is out of scope for
        this OpenAPI document — clients connect to the returned
        `websocket_url` directly.

        Three protocol flavors:
        - `vnc` — RFB 3.x stream from QEMU's VNC server (browser
          consumers);
        - `spice` — SPICE protocol (planned; ROADMAP Backlog);
        - `serial` — raw bytes from QEMU `-serial unix:` chardev
          (CLI consumer `otherix vm console`).

        Tokens are bound to the issuing user, the target VM, and a
        short TTL (30s for the connect step; the websocket session
        itself can run for longer once established).
        Each call issues a fresh token; previous tokens for the same
        VM are not revoked.

        Per-VM single-session lock: an open WebSocket session blocks
        further upgrades with 409 `console_in_use` until the existing
        session closes.
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                protocol:
                  type: string
                  enum: [vnc, spice, serial]
                  default: vnc
                  description: |
                    Console protocol. The agent must support the
                    requested protocol on this VM; if not, 409 is
                    returned with `details.reason = "protocol_not_available"`.
                    `serial` is the CLI-flavored protocol.
      responses:
        "200":
          description: Console session token issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VMConsoleResponse"
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/PermissionDenied" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: |
            VM is not running (`vm_not_running`), the requested protocol
            is not available on this VM (`protocol_not_available`),
            another console session is already open for this VM
            (`console_in_use`), or idempotency-key mismatch
            (`idempotency_key_mismatch`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500": { $ref: "#/components/responses/Internal" }
...
  /v1/tasks:
    get:
      operationId: tasks.list
      tags: [tasks]
      summary: List async tasks visible to the caller.
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, running, success, failed, cancelled]
        - name: type
          in: query
          required: false
          schema: { type: string }
          description: Filter by task type (e.g. `storage_pool.scan`).
        - name: resource_type
          in: query
          required: false
          schema: { type: string }
        - name: resource_id
          in: query
          required: false
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: List of tasks.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskList"
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "500": { $ref: "#/components/responses/Internal" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by `/v1/auth/login` or `/v1/auth/refresh`. Claims:
          - `sub`  — user UUID.
          - `role` — one of `admin`, `operator`, `developer`, `viewer`.
                     The Control Plane resolves the (role, permission, scope)
                     matrix on every request; see `docs/rbac.md`.
          - `iat`, `exp` — standard.

        Role changes take effect on the next access-token refresh; the
        15-minute access-token TTL bounds propagation latency.
    apiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: OtherixAPIKey
      description: |
        Opaque personal API token with `otx_` prefix. Submitted via the same
        `Authorization: Bearer ...` header as JWTs. Server distinguishes by
        prefix.
    agentMTLS:
      type: mutualTLS
      description: |
        Mutual TLS authentication for node agents. The client certificate
        must chain to the cluster CA and carry a SAN of the form
        `cn=node-<name>.agents.otherix.local`. The Control Plane verifies
        both the chain and that the SAN matches a registered node
        (`agent_certs.fingerprint_sha256` joined to `nodes.id`).

        Used for the `agents.*` operations only. User-facing endpoints
        do not accept this scheme; agent endpoints do not accept
        `bearerAuth` / `apiKeyAuth`. Bearer tokens never grant access
        to this surface.
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              $ref: "#/components/schemas/ErrorCode"
            message:
              type: string
              description: Human-readable message in en-US.
            details:
              type: object
              additionalProperties: true
              description: Code-specific structured payload. Optional.
...
    StoragePoolList:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/StoragePool"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
...
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        maxLength: 255
      description: |
        Optional idempotency key. Same key + same body within 24 h returns the
        original cached response. Same key + different body within 24 h returns
        409 `idempotency_key_mismatch`.

    # Name-only path parameters. Each resource family
    # (nodes / vms) accepts a case-insensitive name in
    # the `{id}` path slot and rejects UUID format with 400
    # `validation_failed`. Storage pools keep the polymorphic
    # legacy `{id}` shape per the multi-instance carve-out
    # (per-instance addressing requires UUID disambiguation).
...
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
      description: Page size. Clamped to 200.
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
      description: |
        Opaque cursor returned in the previous response's `meta.next_cursor`.
        Omit on the first page.
...
    NodeIdFilter:
      name: node_id
      in: query
      required: false
      schema:
        type: string
        format: uuid
      description: Filter by node id.
...
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthenticated:
      description: Missing, invalid, expired, or revoked credentials.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    PermissionDenied:
      description: Authenticated but not authorized for this operation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Internal:
      description: Unhandled server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"